Handle lambdas within multi-line expressions in check failures.
Currently if a lambda occurs in a multi-line expression not on the
last line then it will not be added to the cache of nodes for its
file, so when a check failure includes a frame that is a call to the
lambda, no code (or varmap for the innermost frame) for the lambda
will be present.

Now lambdas will be added to the cache of nodes if their last line
number is different than the containing expression. This means that
lambdas will be treated differently depending on whether they are in
the last line of their top-level expression or not: only the lambda
will be displayed for lambdas not on the last line and the entire
containing expression for lambdas on the last line.

Bug: 944848
Change-Id: I95dc56e00c3152130f64cd91abbc5bdeca4e5bce
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1535708
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>