Update the ignores for checkers to include the steps dictionary.
The raw OrderedDict is never passed to check functions, it is instead
copied into a StepsDict. There's no need to pass in the raw
OrderedDict to be ignored and the StepsDict cannot be passed in the
Checker's constructor since it requires the checker itself. Instead of
taking ignores in the constructor, add a method that will update the
ignore set and StepsDict will add itself to the Checker's ignore set.

This will prevent the steps dictionary from being dumped when a lambda
is used as a post process check.

Bug: 944848
Change-Id: Id86f9b18ec82707fe3533c14931a48d4cda31d4d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1535712
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>