Add the ability to use lambdas as post-processing hooks.
Currently, lambdas can't easily be used as post processing hooks because
the return value of check is a bool and the return value of post
processing hooks must either be None or a subset of the input steps
dictionary.

This change adds post_check to RecipeTestApi. It works similarly to
post_process except that the return value of the post processing hook is
ignored, so lambdas that simply call check can be used.

Bug: 944848
Change-Id: Icf9ebc01c283cea5a705565a12849924636589ad
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1535711
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>