Trigger multiple builds from a single run.
The purpose of this recipe is to trigger multiple builds at the same
revision. Specify build request arguments in a property and trigger the
builds in a batch, this may be in separate projects and/or buckets as
long as the service account this recipe runs under has the corresponding
permissions in buildbucket.

R=nodir,liaoyuke,sajjadm,iannucci,phosek

Bug: 944752
Change-Id: Icd2543cc8823ec2dc5b6e4bc80322df0c359683e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1537739
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>