Update post_process_test.py to call magic_check_fn.post_process.
Manually constructing a checker and invoking post process hooks requires
changes to the test whenever the arguments for the checker or the types
expected to be passed to a hook change. Using post_process instead
limits the amount of changes that require updating all of the tests in
post_process_test.py to only those that change the public interface of
the magic_check_fn module.

Change-Id: I66f60cacf5be735cb8d525f5c3fdfc5e7d41c901
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1577825
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>