Convert KeyErrors when executing post-process hooks into check failures.
This provides more helpful information when a KeyError occurs because
the check failure will include information about what keys were in the
dictionary and relevant variables to provide information about what
dictionary was accessed.

This is helpful for creating a step object so that attempting to access
missing keys on dict-fields of the step produces helpful output
(attempting to provide the relevant state as part of the implementation
of the fields was drastically increasing the complexity). It also removes
the need for the StepsDict type to get helpful error messages when
attempting to access a missing step.

The code for checking for KeyErrors when producing variable information
enables the use of common idioms involving short-circuiting logical
operators: "check(key in d and d[key] == some_value)" and
"check(key not in d or d[key] != value)".

Change-Id: I4470a42520d1ceb82bdba2e0f472ddf22d204c32
Bug: 939120, 946015
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1565191
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>