Add an object for representing a step.
The Step class has fields for all of the potential keys that would be in
a step dict except for ~followup_annotations. Instead of having a field
for ~followup_annotations, Step has a field for all of the pieces of
information that were contained in ~followup_annotations.

To temporarily maintain backwards-compatibility, Step implements the
collections.Mapping interface. Modifications to a Step field is reflected
in the value of corresponding key, whether the key field corresponds to
its own key in the dictionary or an annotation in ~followup_annotations.

Change-Id: Iccf636ff5f458e145bc13ba843d66e06c61accc2
Bug: 939120
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1542435
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>