Fix masked ImportErrors for config.py and test_api.py.
Previously we imported these with

  try:
    ...import
  except ImportError:
    pass # assume module doesn't exist.

However, this is wrong if the imported module actually raises
ImportError (because it has a bad import). In this case we would
silently swallow the import error and pretend like config and/or
test_api just don't exist.

Now we just do a file existence check for config.py and
test_api.py and let the ImportErrors bubble up to the user.

Bug: 961418
Change-Id: I4af61b68b66bceeafb6aec5e7be505e118f81e56
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1603829
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Sean Abraham <seanabraham@chromium.org>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>