Stringify recipe_name when executing test case.
Because this was previously done at a weird place in
the old test runner. The raw recipe name is actually unicode,
and the logic for executing the recipe (which manipulates the
properties it passes into the recipe) is a more natural place
to do this.

R=tandrii@chromium.org

Bug: 910369
Change-Id: Ib5ba28270bc0a8a782ed8016bb19772442a970b3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1617301
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>