Check validity of _resolved nodes before accessing them.
Post processing expressions that result in KeyErrors can result in
exceptions producing the check failure information if the expression
resulting in a KeyError is part of a larger expression. Checking the
validity of a _resolved node prevents that.

Change-Id: I617112914a88db6339777b30aad13b6d92b09a61
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1623539
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>