Fix timeout bug in subproc runner.
When a timeout occurs we actually need to kill the process group;
this was lost in a previous refactor, even though we have a test
for it (!!). The upcoming gevent runner actually correctly hangs
on this test; not sure what went wrong here.

Additionally s/has_timeout/had_timeout/ to avoid secondary
exception.

TBR=tandrii@chromium.org

Change-Id: I6b7dc5f51399f694f82393e8321a078a39a2a7f6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1626130
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>