Implement new test runner.
This caps off the last couple CLs and implements a new test runner. In
particular, this removes the need for multiprocessing, which paves the
way to introduce gevent (finally).

The new test runnner seems to be roughly the same performance as the
existing test runner (on build.git).

The error reporting pipeline for this version is much clearer, and it
should be trivial to add additional checks / detail for users.

R=tandrii@chromium.org

Bug: 910369
Change-Id: I55ea985093b2eabdbbe7f672432069fa872e7a31
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1619427
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>