Remove ability to set callback on nesting steps.
Nothing got the chance to use this (hopefully).
Also beef up the documentation around `api.step.nest`.

This makes it so that the nest function just directly yields
the StepPresentation... which is the only object that anyone
would care about anyhow :).

Unfortunately, because all the existing code expects a
StepData-containing-a-StepPresentation to be yielded, we have
to add an awkward proxy object around the presentation to
avoid needing to update all code in the ecosystem.

R=tandrii@chromium.org

Bug: 910369
Change-Id: I26f27d9499e43cbb532f9b176b8a9dd6d7f2251c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1629438
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>