Move tracking of namespaces and deduplicated step names into the engine.
Previously this was the combined responsibilities of the 'context' and
'step' modules. Now it's entirely the responsibility of the engine.

This will allow the recipe engine to provide cohesive state tracking for
greenlets (since they need to copy from and synchronize with the global
step name information).

R=tandrii@chromium.org

Bug: 910369
Change-Id: I18bf4b3a24d1b4c99231e37b9328733c74d1f579
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1635601
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>