Get rid of gevent monkey patching.
Since we control all of the code which needs to be made cooperative,
there's no benefit to doing the monkeypatch.

Additionally, the logdog butler client seems to be incompatible with the
gevented `socket` library. I suspect it may be some combination of
{domain sockets + OS X + gevent}, but regardless we always expect butler
client library writes to be ultrafast, so making them nonblocking doesn't
actually buy anything.

If it turns out in the future that we NEED to make it nonblocking, we
can explicitly import `gevent.socket` there too (after we figure out and
fix the bugs).

R=tandrii@chromium.org

Bug: 910369
Change-Id: I3d7393400969c5c47b6b913ed66b09efd60ff2d7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1638250
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>