Add cpu cost classes to api.step.
This allows recipes to restrict the number of steps which are actually
running simultaneously by contending on CPU as a finite resource.

Note that this does not restrict the number of greenlets that the
recipe can run simultaneously. The intent is that the recipe can
start an unbounded number of greenlets in a natural style (e.g.
fan out over 100 steps to run), but prevent 100 subprocesses
from actually fighting each other over the CPU (e.g. with 100
unittests).


R=tandrii@chromium.org

Bug: 910369
Change-Id: I96fb7758f0d6066cceaca6a75e5e439390b478d0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1639585
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>