Safely check _finalized in StepData.__setattr__.
The current code uses hasattr to check the value of _finalized in
__setattr__ but hasattr calls getattr and checks whether an exception is
raised. This causes infinite recursion when fields are being set during
initialization because __setattr__ is called, which causes the value of
_finalized to be checked, which causes __getattr__ to be called for an
attribute that is not yet on the object, which causes self.name to be
accessed to provide a better error message, which accesses self.name,
which in turn accesses self.name_tokens, which causes __getattr__ to be
called for name_tokens which is not yet set on the object.

Accessing the object's __dict__ directly avoids any additional dunder
methods being called, preventing the infinite recursion.

Bug: 965675
Change-Id: I7211f4f7b27ed11eae5cd22bade4586667a57df2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1677375
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>