Safely check _finalized in _AttributeRaiser.__setattr__.
The current code uses hasattr to check the value of _finalized in
__setattr__ but hasattr calls getattr and checks whether an exception is
raised. This causes infinite recursion when fields are being set during
initialization because __setattr__ is called, which causes the value of
_finalized to be checked, which causes __getattr__ to be called for an
attribute that is not yet on the object, which causes self._step_name
and self._namespace to be accessed to provide a better error message,
which causes __getattr__ to be called for those attributes that are not
yet set on the object.

Accessing the object's __dict__ directly avoids any additional dunder
methods being called, preventing the infinite recursion.

Bug: 965675
Change-Id: I60155839b1548816da472dd8743111b9917dce77
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1678175
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>