[commands/test] Use "." instead of checkmark for success, even with emoji.
This reduces the amount of visual noise, since 'success' is the vast
majority of statuses. This allows other emoji (like 💾) to be easier to
see in the output.

Adjusts the printing code so that it uses column counting (instead of
symbol counting), and this counting only engages when stdout is not a
tty (e.g. on bots).

R=martiniss@chromium.org, tandrii@chromium.org, thakis@chromium.org

Bug: 985349
Change-Id: Ic167d40e766199ba598c720230f51e0f623a903f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1708216
Commit-Queue: Andrii Shyshkalov <tandrii@chromium.org>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>