[swarming] Make TaskResult.output_dir a Path
Uses output_dir.join(...) instead of api.path.join(output_dir, ...) to
construct the swarming TaskResult output_dir. This makes the output_dir
a Path rather than a string, for better compatibility with recipes that
use the swarming module.

This change is technically not backward compatible, because the
output_dir argument to api.swarming.collect() can now only be a Path,
not a string. However, the docstring already indicated that it only
accepted Paths. And based on a quick code search, there doesn't seem to
be any recipe code that passes a string as the output_dir, so this
should be safe.

I also renamed the `id` variable to fix pylint's complaint that it
conflicts with the id() built-in.

Test: https://ci.chromium.org/swarming/task/46446502767d6710?server=chromium-swarm.appspot.com
Change-Id: I83d4ee157f6f264afcc3547d0b376f24673bd3b4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1719380
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>