Account for raw_io output backing file not existing when reading it.
When using the leak_to argument for the raw_io output methods, a
temporary backing file isn't created and instead the command that the
returned PlaceHolder is passed to is expected to create the file at that
location. If the command did not create the file, this leads to an
exception trying to read the file to get the result data to add to the
step result.

This fix accounts for this case by silencing the exception when reading
the file if the file does not exist.

Bug: 930963
Change-Id: I2a7990b35c1ad0ab98a5abdf5a76d46c22b6232f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1733096
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>