[led] Create LedTestApi
Adds a LedTestApi for mocking test steps that make led calls.
Constructing realistic mock return values for led commands is somewhat
tricky because the led commands generally are piped into each other to
make incremental modifications to the same base input (which is normally
the output of `led get-build` or `led get-builder`).

The way I achieved this is by storing the builder data in a
`LedTestData` object and modifying it (along with mocking the
corresponding led command's return value with the modified data)
whenever a method is called on the `LedTestData` object. It's not
particularly pretty, but it works.

Fuchsia-Bug: https://bugs.fuchsia.dev/p/fuchsia/issues/detail?id=34809
Change-Id: Ibc8b4aa273c497536ec2dd598641f4f1bb9739f8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1757083
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Oliver Newman <olivernewman@google.com>