[tricium] Type check comments early
The fuchsia tricium analyzer recipe seems to be calling this function
with strings in some cases, which is leading to errors on the tricium
server when it tries to unmarshall the object. It would be much easier
to debug if the recipe failed rather than succeeded.

Bug: 990100
Change-Id: Ib4ecd0226e9907a0af33688dab10d0767dea2933
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1783840
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Auto-Submit: Gary Miguel <garymm@google.com>