[testing] Have @placeholder_step_data take a name
Adds an optional function name to the @placeholder_step_data decorator.
If provided, the data returned by the wrapped RecipeTestApi method will
be associated with the named placeholder method - rather than the
placeholder method with the same name as the wrapped method - on the
corresponding RecipeApi. The name defaults to the name of the wrapped
RecipeTestApi method, so it's backwards-compatible.

This will allow creation of more methods like JsonTestApi.invalid()
without duplication of the name-overriding logic.

Change-Id: I097eb9e4bf2da683eae6828890c38997e5fd6710
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1788034
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Oliver Newman <olivernewman@google.com>