[raw_io] Make missing placeholder files testable
If an `OutputTextPlaceholder` backing file doesn't exist when `result()`
is called, `result()` will return None. Most calling code does not
handle this case, in part because the testing-specific code in
`OutputTextPlaceholder.result()` didn't have a way to mock a missing
backing file. This caused occasional InfraFailures in downstream repos
that assumed that the return value of `result()` was non-null.

This change adds support for mocking a missing backing file by passing
`None` as the data to `RawIOTestApi.output()`. This instructs
`OutputTextPlaceHolder.result()` to return None, as it does when the
backing file is missing. This will let downstream repos add tests for
the missing backing file case.

For maximal backwards compatibility, a missing backing file can only be
simulated if the `leak_to` argument to the placeholder is unset. There
are a lot of steps in downstream repos that use output placeholders but
don't mock the placeholder result in the tests. In these cases, the
placeholder used to evaluate to an empty string, but now evaluates to
None. This would break a lot of code that assumes the result is a
string. To avoid this, when in testing mode, the placeholder will
default to the old value of returning an empty string *unless* leak_to
is specified (which is rare), in which case it will return None.  This
is reasonable because it's entirely possible that the `leak_to` file
path won't exist.  On the other hand, it's impossible that the temporary
file created if `leak_to` is unset would be missing, unless the recipe
explicitly deletes it before `result()` gets called.

Bug: fuchsia:35478
Change-Id: Ia4312e157945a18a3f81c33789f41a405e1b6dfa
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1777143
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>