[buildbucket] Let child builds use builder priority
Recipes can now pass `priority=None` to
`api.buildbucket.schedule_request` to instruct Buildbucket to use the
destination builder's priority, as defined in the configs, as the
priority of the child build.

This is mostly backward compatible: if the `priority` argument is unset
it will still default to the priority of the parent build. The only
exception is the case where downstream code explicitly sets `priority`
to a falsey value. This was a no-op previously since `None` was the
default, causing the child build to have the same priority as the
parent. Now, any falsey value will give the child build the priority
defined by its builder configs.

In practice no downstream repos currently use a falsey priority, so this
will not break anything downstream.

I also added a `priority` argument to all the `*_build_message()`
buildbucket test_api methods so that recipe tests (including the tests
for this change) can mock the priority of the build.

Bug: fuchsia:38153
Recipe-Nontrivial-Roll: build
Recipe-Nontrivial-Roll: build_limited_scripts_slave
Recipe-Nontrivial-Roll: chromiumos
Recipe-Nontrivial-Roll: fuchsia
Change-Id: Idbc679b0b53c0869751f28c55d56e630a733dc04
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1846716
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>