[buildbucket] Inherit child fields from parent
Changes the default value for most other
`api.buildbucket.schedule_request()` arguments to be the `INHERIT`
sentinel orginally created for use as the `priority` argument, which
instructs the method to use the value of the parent build for that
argument. This makes it more explicit that these arguments' values
default to those of the parent build. (Following up on
http://crrev.com/c/1846716)

The `project` parameter requires a soft transition because of this
method in the "build" downstream project:
https://chromium.git.corp.google.com/chromium/tools/build/+/d467035c66eb98ebdd829be9bb07078136ad17ea/scripts/slave/recipe_modules/v8/api.py#1448
which can set `None` as the value of the `project` argument. I'll submit
a downstream CL to have it use `INHERIT` instead.

All expectation diffs are a result of explicitly setting the child
gitiles_commit even if empty, and fixing buggy behavior regarding the
`experimental` argument. `experimental` defaults to the value of the
parent's build's `experimental` field, which is a boolean, but it gets
passed to the `experimental` field of `ScheduleBuildRequest`, which is a
Trinary.  Previously, if the parent build had `experimental=False`, the
child build request would get `experimental=Trinary.UNSET` (the protobuf
default). The build request will now instead get
`experimental=Trinary.NO`, and any callers that want to have
`experimental=Trinary.UNSET` will have to pass it explicitly.

Recipe-Nontrivial-Roll: build
Recipe-Nontrivial-Roll: build_limited_scripts_slave
Recipe-Nontrivial-Roll: chromiumos
Recipe-Nontrivial-Roll: fuchsia
Recipe-Nontrivial-Roll: release_scripts
Bug: fuchsia:38153
Change-Id: I05a7da97bb887be5aa8ae9210aabb538df6a9a60
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1854019
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>