[buildbucket] Don't accept project=None in schedule_request
This CL:
https://chromium-review.googlesource.com/c/chromium/tools/build/+/1859576
got rid of the only downstream usage of `project=None` defaulting to the
parent build's project. So it's now safe to no longer support that, and
have only `project=INHERIT` cause `schedule_request()` to default to the
parent's project.

This is a follow-up to
https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1863249.

Change-Id: Icbaa42fab63a77ca3e829499c5e732a38c93ea9d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1865402
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>