[recipe_modules/context] Simplify context implementation.
It turns out that having a list of values is 100% unnecessary. We used
to do this because we stored all the intermediate values in the list
and then assembled them at retrieval time, but at the time of this CL
we were just putting the full current value on the end of the list.
The python stack is plenty-good for this purpose, too...

R=tandrii@chromium.org

Bug: 1026340
Change-Id: I983cad8adc1bab1fa185c64ddb6e4a5515175bfc
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/1937428
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>