Report run steps in the order they were run.
For a test failure, the steps that were run are reported to the user,
but they are in an arbitrary order. Instead, order them according to the
order the steps were run so that the used can more easily find where an
error has occurred.

Change-Id: I5b38e61401a7cd93d9ec47db5abf8944ee0235ae
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/2055364
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>