[luciexe] Fix the issue where build status is not properly updated
If error is raised during recipe execution, it is currently caught
within the engine and the error message is stored in the result proto.
However, currently, luciexe is trying to catch the error at top level
and update the build status according to the exception type, therefore,
the build status will almost always be `SUCCESS` as no error will be
caught at this layer.

This change fixs the issue by updating the build with the recipe excution
result after finish running the recipe.

R=iannucci

Bug: 1068432
Change-Id: Ie844091fdc3f2a253cb72895f51af0277de72dbb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/2138122
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>