Always set gevent tracing during simulations.
Now that more and more recipes are going to use concurrency features,
effectively every recipe needs gevent tracing to function. Bite the
bullet and set it for all simulations.

This will make simulation tests slower. On my 2014 Macbook Pro
(Core i7, 2.8GHz, 4 core/8 thread), this makes build.git take 140s up
from 90s for a `test train` (+50%). I expect this to be less bad on
higher-powered machines, but it's still pretty bad. Still, I think the
tradeoff is worth it; This is about as bad as it will get, and this will
now actually be correct (instead of maddeningly reporting arbitrary
chunks of the recipe code as being uncovered, even when they CLEARLY
are).

R=olivernewman@google.com, tandrii@chromium.org, vadimsh@chromium.org, yiwzhang@google.com

Bug: 1015181
Change-Id: I071362e7c71ef07802af88a71a73badc225951b7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/2223350
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>