Add swarming.ensure_client().
In order to be able to get the swarming client ready in parallel with
performing operations leading up to its use, there should be a public
method that can be called that results in the client being fetched.
Currently, all of the public methods either do not cause the client to
be fetched, actually use the client or is a context manager, which is
awkward to trigger a greenlet for.

Change-Id: I18e28076448f657e3860133c6b3dbbd701bbbf60
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/2230780
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Commit-Queue: Takuto Ikuta <tikuta@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>