[post_process] Fix ResultReasonRE
ResultReasonRE wasn't properly validating that the recipe result reason
contained the specified text. This is because the second argument to
check() is optional and previously defaulted to None. If it was set to
None, the function assumed it was unset and considered the *first*
argument to be the condition it should check.

This worked in most cases, but not for ResultReasonRE, because
re.match() returns None if the string doesn't match the regex. This
caused check() to check the first argument instead, which was the
non-empty hint string. Since that string is truthy, the check always
passed:
https://github.git.corp.google.com/luci/recipes-py/+/5c575266cf91a14adf9b17c71f4e4e03ad964c24/recipe_engine/post_process.py#505

I fixed this by having the argument default to the `MISSING` sentinel
instead of None, which lets check() distinguish between the argument
being unset vs. being set to None.

I also could have fixed this by bool()ing the return value of re.match()
in ResultReasonRE, but fixing the underlying check() implementation
seemed like the better way to go to prevent other people hitting this
issue in the future.

Recipe-Manual-Change: build_limited_scripts_slave
Recipe-Manual-Change: chromiumos
Change-Id: I33c7475aa24008dd8dcdb15aed46d26522e5a1fc
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/2274730
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>