[resultdb] use : instead of = as a delimiter for -var
In https://crrev.com/c/2435049, the delimiter for -var was changed from
'=' to ':' to use the same delimiter with -tag.

The CL also made a temporal patch to `rdb stream` such that it will
support both delimiters for a short period. Once this CL is submitted
and the change is deployed to all builders, the temporary patch from
`rdb stream` will be removed.

R=chanli@chromium.org,nodir@chromium.org,iannucci@chromium.org

Bug: 1099573
Change-Id: I31af35c733dc4b7d5b349233364855e79fad9e3c

Recipe-Nontrivial-Roll: fuchsia
Recipe-Manual-Change: build
Change-Id: I31af35c733dc4b7d5b349233364855e79fad9e3c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/2437000
Commit-Queue: Scott Lee <ddoman@chromium.org>
Reviewed-by: Scott Lee <ddoman@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>