[luci_context] Move writing of LUCI_CONTEXT files to the engine core.
This removes an odd file-write from 'userspace' recipe code, but also
enables compuation of 'timeout' for the step much closer to the actual
step execution. Specifically, this avoids having 'timeout' also cover
the time that we block for resource contention when running multiple
greenlets in the recipe.

R=vadimsh, yiwzhang

Bug: 1127089
Change-Id: Iff59e7698e9556737a6144508c00cb3342811641
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/2441885
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>