[cipd] Handle empty responses from CIPD
During a recent GCS outage, we saw some `cipd ensure` failures where
CIPD didn't produce any JSON output. Then the recipe would crash with a
`TypeError("argument of type 'NoneType' is not iterable")` trying to
look up the 'error' key in the JSON output. For example:
https://ci.chromium.org/b/8865717012587636816

To avoid this, skip checking for the 'error' key if the JSON output is
falsy and just re-raise the original StepFailure.

Bug: fuchsia:62641
Change-Id: I2bcd9256e5142dae602728a71feea7fba93e4736
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/2495134
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Oliver Newman <olivernewman@google.com>