[path] Avoid unnecessary temp dir renamings in tests
Previously, we kept track of a single global `_test_counter` integer
that was incremented on every call to `mkstemp()` or `mkdtemp()`, and
then the new value was used in the mock temporary file name to ensure
uniqueness. The problem with this is that it led to unnecessary
expectation file changes when adding new `mkdtemp()` or `mkstemp()`
calls to recipes, because doing so would change the names of all
temporary files/directories that were created later in the recipe.

To fix this, keep track of the number of temporary files we've created
for each `prefix` argument (which is the bare minimum necessary to
ensure uniqueness) instead of across all prefixes. This will make it far
less likely that a new `mkdtemp()` or `mkstemp()` call will change the
names of later temporary files, as long as every usage of those
functions in the code uses a different prefix (which is often the case
anyway).

This will use a bit of extra memory: O(number of prefixes used) instead
of O(1). However, this shouldn't be too much of a concern because it
only applies to unit test runs, not to production recipe runs.

Recipe-Nontrivial-Roll: build
Recipe-Nontrivial-Roll: build_limited_scripts_slave
Recipe-Nontrivial-Roll: chrome_release
Recipe-Nontrivial-Roll: chromiumos
Recipe-Nontrivial-Roll: depot_tools
Recipe-Nontrivial-Roll: fuchsia
Recipe-Nontrivial-Roll: infra
Change-Id: I6e4c8e69ca4dd64a45a53e93ccd965000fd7b7db
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/2514557
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>