[luciexe] stop using FileObjectPOSIX when sending data to logdog butler
It looks like sometimes data is not sent to lodgog butler if the client (
i.e. a socket) is wrapped with FileObjectPOSIX. It could be a bug in the
gevent libray or I use it in a wrong way.

This is causing Build page to be blanket while recipe is running fine.
Therefore, ditch this feature completely because:
  * Future module (parallel recipe) is not widely used so this feature
    won't give us much performence gain.
  * It is currently only enabled on POSIX, which means we will anyway
    need to figure out a general solution that works for both platforms
    in the future.
  * If we implement go/butler-protocol-v2, it's likely we don't need it
    at all because sending log become a plain http request.


R=tandrii

Bug: 1150172
Change-Id: I9300ed226f4e43441efaa35c5eaa0215785e4ff3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/2546059
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>