Return None from CommitList.advance_to for backwards advances.
Prior to this change, when a call to advance_to is made with a
target_commit that is before the commit list's current position, the
position is left untouched and the commit at that current position is
returned.

This exacerbates issues that cause the roller to sometimes get stuck.
The candidate algorithm intends to prevent backwards rolls, but since
they are not correctly detected, it continues advancing dependencies and
the recipes cfg proto will be updated with the incorrect revision values
where the backwards rolls were attempted and the set of dependencies
will be found inconsistent. If they were correctly detected as backwards
rolls, fewer dependencies would be advanced and therefore fewer
potential candidates prematurely eliminated.

With this change, calls to advance_to with with a target_commit that is
before the commit list's current position will return None, indicating
that the advance_to operation was not able to proceed. This will result
in fewer dependencies being erroneously advanced, which will increase
the number of potential candidates that can be considered.

Bug: 1171386
Change-Id: Ibeeb9a1e55885599dac00e3fedb59e061c2eab99
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/2653173
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>